python-service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
httpx
pydantic
pydantic-settings
python-dotenv
